+++ examples/gjs/gio-2-cat-packages/@types/Gjs/gio-2.0/gio-2.0.js
@@ -1,0 +1,8 @@
+
+
+// @ts-expect-error
+import Gio from 'gi://Gio?version=2.0';
+export { Gio };
+export default Gio;
+
+