+++ examples/node-gtk/gtk-3-editor/main.ts
@@ -33,8 +33,5 @@ wnd.connect('delete-event', () => {
 })
 
 console.log('hello world')
-// import { gettext } from '@types/Gjs/Gjs';
-// let locale = gettext.setlocale(gettext.LocaleCategory.ALL, null);
-// console.log('locale', locale);
 
 Gtk.main()