+++ packages/lib/src/types/generate-config.ts
@@ -7,7 +7,7 @@ export interface GenerateConfig {
     /** javascript environment */
     environment: Environment
     /** root / working directory of your project */
-    root: string;
+    root: string
     /** directory to output to */
     outdir: string | null
     /** Definitions generation type */